+++ .github/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ### New features since last release
 
+ * Add support for building X86-64 Linux wheels with OpenMP and SERIAL backends with Github Actions.
+ [(#14)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/14)
+
+ * Add the `kokkos_config` class variable, which stores the kokkos build and runtime information such as `Backend`, `Architecture`, `Kokkos Version`, `Compiler`, to LightningKokkos for users' query purpose. Users can also access other information such as `Options`, `Memory`, `Atomics` and `Vectorization` from `kokkos_config`.   
+ The workflow for build and runtime information query is:
+ ```python
+ >>> import pennylane as qml
+ >>> dev = qml.device('lightning.kokkos', wires=3)
+ >>> dev.kokkos_config["Backend"]
+ {'Parallel': 'OpenMP'}
+ ```
+ [(#14)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/14)
+
 ### Breaking changes
 
 ### Improvements
@@ -70,7 +83,7 @@ Amintor Dusko, Lee J. O'Riordan, Shuli Shu, Matthew Silverman
  [(#11)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/11)
 
  * Add sample generation support.
-  [(#9)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/9)
+ [(#9)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/9)
 
 ### Breaking changes